feature: make GoodsFilterMain component
