+++ src/components/Filters/GoodsFilters/GoodsFilterMain/style.module.scss
@@ -1,0 +1,51 @@
+.filter {
+  display: flex;
+  flex-direction: column;
+  gap: 36px;
+  width: 290px;
+  &__fieldset {
+    display: flex;
+    flex-direction: column;
+    gap: 21px;
+    border: none;
+    padding: 0;
+    margin: 0;
+  }
+  &__legend {
+    margin: 0 0 28px;
+    font-size: 20px;
+    font-weight: 700;
+    letter-spacing: .6px;
+  }
+  &__link {
+    width: fit-content;
+    text-decoration: none;
+    color: inherit;
+    margin: 0;
+    font-size: 16px;
+    font-weight: 500;
+    line-height: 20px;
+    &:hover {
+      color: var(--link-color, #9384B7)
+    }
+    &_active {
+      color: var(--purple-pushed-color, #615779)
+    }
+  }
+  &__add-btn {
+    border: none;
+    padding: 0;
+    background-color: transparent;
+    text-align: left;
+    width: fit-content;
+    font-size: 20px;
+    font-weight: 500;
+    color: var(--link-color, #9384B7);
+    transition: opacity .3s ease;
+    &:hover {
+      cursor: pointer;
+      opacity: .8;
+    }
+  }
+
+}
